% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shiny.R
\name{applyFilter}
\alias{applyFilter}
\title{Add TRUTH column to cdata}
\usage{
applyFilter(
  cdataDF,
  filterDF,
  cell_unique_id_field = "ucid",
  truth_column = "filter"
)
}
\arguments{
\item{cdataDF}{The \code{cdata} data.frame.}

\item{filterDF}{The filtering data.frame as produced by \code{\link{calculateTruth}} in \code{\link{polyFilterApply}}.}

\item{cell_unique_id_field}{Name for the column holding the unique identifier (a "primary key") for each data point (i.e. the "ucid" is not suficcient for time series datasets).}

\item{truth_column}{Name of the column in \code{filterDF} with the boolean filter.}
}
\description{
This function is meant to be called exclusively by \code{\link{polyFilterApply}},
as it produces appropriate inputs.
}
\keyword{internal}
