% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shiny.R
\name{calculateTruth}
\alias{calculateTruth}
\title{Calculate how filters apply to data according to the logic "mode"}
\usage{
calculateTruth(
  filterDF,
  cell_unique_id_field = "ucid",
  truth_column = "filter",
  mode = "all",
  inclusive_logic_name = "Include",
  exclusive_logic_name = "Exclude"
)
}
\arguments{
\item{filterDF}{The "cfilter" dataframe, as built by \code{\link{polyFilterApply}} by iteratively calling \code{\link{polyFilterCell}}.}

\item{truth_column}{Name of the column with the boolean filter.}

\item{mode}{Filter logic priority. Must be either "all" (exclusion overrides inclusion) or "any" (inclusion overrides exclusion).}

\item{inclusive_logic_name}{Logic name for the "inclusive" group of filters, selecting their union / inclusive or.}

\item{exclusive_logic_name}{Logic name for the "exclusive" group of filters, selecting the complement of their union.}
}
\description{
This function is meant to be called exclusively by \code{\link{polyFilterApply}},
as it produces appropriate inputs.
}
\keyword{internal}
