% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{cellSpreadPlot}
\alias{cellSpreadPlot}
\title{Plot for 2D binning of data and tiling of cell pictures}
\usage{
cellSpreadPlot(
  cdata,
  paths,
  ch = "BF.out",
  boxSize = 80,
  xvar = "a.tot",
  yvar = "fft.stat",
  overlay_points = F,
  underlay_points = F,
  draw_contour_breaks = NULL,
  x.cuts = 7,
  y.cuts = 7,
  ...,
  discrete_exp = 0.01
)
}
\arguments{
\item{cdata}{A "cdata" data.frame with the CellID variables (not the cell.data object, \code{cell.data$data}). It can be obtained from \code{cell.load.alt} or \code{load_cell_data}.}

\item{paths}{A "paths" data.frame, with paths to the images (i.e. \code{cell.data$images}), and t.frame, position and channel information of each one. It can be obtained from \code{cell.load.alt} or \code{load_cell_data}.}

\item{ch}{Name of the CellID channel (BF, BF.out, RFP, etc.). "BF.out" by default, use a vector to select more than one channel simultaneously (images will be stacked).}

\item{boxSize}{Lateral size of the square box extracted from the original images (in pixels). 50 px by default.}

\item{xvar, yvar}{Strings indicating names for the variables to plot in the horizontal (x) and vertical (y) axis.}

\item{overlay_points}{Overlay data points to the image plot.}

\item{underlay_points}{Underlay data points to the image plot.}

\item{draw_contour_breaks}{Overlay a ggplot2::stat_density2d layer. If TRUE, use the default breaks. Otherwise NULL for none, or a numeric vector for the density breaks.}

\item{x.cuts, y.cuts}{Integers indicating the number of cuts for each variable.}

\item{...}{
  Arguments passed on to \code{\link[=cellSpread]{cellSpread}}
  \describe{
    \item{\code{xvar,yvar}}{Strings indicating names for the variables to plot in the horizontal (x) and vertical (y) axis.}
    \item{\code{x.cuts,y.cuts}}{Integers indicating the number of cuts for each variable.}
    \item{\code{for_plotting}}{Return value changes to list of elements important for plotting.}
    \item{\code{cut.by.content}}{Use quantile to generate cuts of "roughly equal content (rather than length)."}
  }}

\item{discrete_exp}{Expansion argument passed to \code{scale_x/y_discrete}, increase to 0.5*1E-2 or more if you see weird artifacts.}
}
\description{
Plot for 2D binning of data and tiling of cell pictures
}
