% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{fill_missing_frames}
\alias{fill_missing_frames}
\title{Fill musing frames using tidyr::fill}
\usage{
fill_missing_frames(cdata.one, min_frame = NULL, max_frame = NULL)
}
\arguments{
\item{min_frame}{Fill cdata up from this frame.}

\item{max_frame}{Fill cdata up to this frame.}
}
\value{
cdata object with minimal columns and filled missing frames.
}
\description{
Fill musing frames using tidyr::fill
}
