% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shiny.R
\name{hexPlotDf}
\alias{hexPlotDf}
\title{Armar data.frame para plots tipo HEXBIN con facets}
\usage{
hexPlotDf(
  cdata,
  facetVars = c("pos", "treatment"),
  varx = "a.tot",
  vary = "el.p"
)
}
\arguments{
\item{cdata}{A data.frame (tipically rcell's \code{cdata}).}

\item{facetVars}{Charachter vector with the names of grouping variables (tipically in rcell's \code{pdata})}

\item{varx}{Name of the x-axis variable.}

\item{vary}{Name of the y-axis variable.}
}
\value{
A data.frame with columns: cID (bin ID), counts, x, y, and "faceting" variables.
}
\description{
Armar data.frame para plots tipo HEXBIN con facets
}
\keyword{internal}
