% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shiny.R
\name{hexPlotDf_test}
\alias{hexPlotDf_test}
\title{Test plot for hexPlotDf}
\usage{
hexPlotDf_test(
  h,
  facetVars = c("pos", "t.frame"),
  varx = "a.tot",
  vary = "el.p"
)
}
\arguments{
\item{h}{The hexPlotDf output.}

\item{facetVars}{Charachter vector with the names of grouping variables (tipically in rcell's \code{pdata})}

\item{varx}{Name of the x-axis variable.}

\item{vary}{Name of the y-axis variable.}
}
\description{
Test plot for hexPlotDf
}
\keyword{internal}
