% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{image_border_one}
\alias{image_border_one}
\title{Add a border to one side of the image}
\usage{
image_border_one(image, geometry = "0x15", color = "white", flip = F)
}
\arguments{
\item{image}{magick image object.}

\item{geometry}{From \code{?image_border}: A geometry string to set height and width of the border, e.g. "10x8". The default value adds a 15px border to the top of the image.}

\item{color}{color string for the border background.}

\item{flip}{Flip the side where the border ends up.}
}
\value{
Image magick object with borders
}
\description{
image_border_one is good for padding images to add captions later, while magick's original image_border adds borders symmetrically to opposing sides.
}
