% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{magickForKnitr}
\alias{magickForKnitr}
\title{Display an image in rmarkdown with knitr}
\usage{
magickForKnitr(
  imgs,
  format = "png",
  .prefix = "tile",
  .resize = NULL,
  .path = tempdir(),
  .file = NULL,
  .include = TRUE,
  .print = FALSE,
  ...
)
}
\arguments{
\item{imgs}{a magick image or cellMagick output}

\item{format}{output format such as \code{"png"}, \code{"jpeg"}, \code{"gif"}, \code{"rgb"} or \code{"rgba"}.}

\item{.prefix}{A string prepended to the file name, "" by default.}

\item{.resize}{a cellMagick image resize string as "200x200" (default NULL, for no resizing).}

\item{.path}{Directory where the output should be saved.}

\item{.file}{Full path to the file name where the image will be saved. Overrides all other file name parameters.}

\item{.include}{Set to true to use knitr::include_graphics directly.}

\item{.print}{Print the image path.}

\item{...}{
  Arguments passed on to \code{\link[magick:editing]{magick::image_write}}
  \describe{
    \item{\code{path}}{a file, url, or raster object or bitmap array}
    \item{\code{density}}{resolution to render pdf or svg}
    \item{\code{depth}}{color depth (either 8 or 16)}
    \item{\code{defines}}{a named character vector with extra options to control reading.
These are the \verb{-define key{=value}} settings in the \href{http://www.imagemagick.org/script/command-line-options.php#define}{command line tool}.
Use an empty string for value-less defines, and NA to unset a define.}
    \item{\code{image}}{magick image object returned by \code{\link[magick:image_read]{image_read()}} or \code{\link[magick:image_graph]{image_graph()}}}
    \item{\code{quality}}{number between 0 and 100 for jpeg quality. Defaults to 75.}
    \item{\code{comment}}{text string added to the image metadata for supported formats}
    \item{\code{flatten}}{should image be flattened before writing? This also replaces
transparency with background color.}
    \item{\code{compression}}{a string with compression type from \link[magick]{compress_types}}
  }}
}
\value{
An path to a temporary image file.
}
\description{
Display an image in rmarkdown with knitr
}
