% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shiny.R
\name{polyFilterCell}
\alias{polyFilterCell}
\title{Build filterDF from polygonDF and cdataDF}
\usage{
polyFilterCell(cdataDF, filterDF, polygonDF, polygonName = 1)
}
\arguments{
\item{cdataDF}{A "cdata" dataframe.}
}
\description{
Build filterDF from polygonDF and cdataDF
}
\keyword{internal}
