% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppServer.R
\name{shinyAppServer}
\alias{shinyAppServer}
\title{Shiny app server function for shinyCell}
\usage{
shinyAppServer(input, output, session)
}
\arguments{
\item{input}{provided by \link{shinyCell}.}

\item{output}{provided by \link{shinyCell}.}

\item{session}{provided by \link{shinyCell}.}
}
\description{
Define server logic
}
\keyword{internal}
