% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagCell.R
\name{updateVector}
\alias{updateVector}
\title{Update a named vector's value using another named vector, by common names.}
\usage{
updateVector(v1, v2, only.common.names = T)
}
\arguments{
\item{v1}{List to be updated (with the "original" or "old" values).}

\item{v2}{List used for updating (with the "newer" values). Note: it needn't have all names.}

\item{only.common.names}{Names of `v2` absent in `v1` will be ignored, unless `only.common.names=F`.}
}
\description{
Names of `v2` present in `v1` will update values in `v1`.
}
\details{
Names of `v2` absent in `v1` will be ignored, unless `only.common.names=F`.
}
