% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{x_to_ranges}
\alias{x_to_ranges}
\title{Make a string of ranges from a vector of integer numbers}
\usage{
x_to_ranges(x)
}
\description{
Make a string of ranges from a vector of integer numbers
}
