% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shiny.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply shiny-filters to cdata}
\usage{
apply_filters(
  cdata,
  filters,
  unique_id_fields = c("ucid", "t.frame"),
  truthMode = "all"
)
}
\arguments{
\item{cdata}{A "cdata" data.frame with the CellID variables (not the cell.data object, \code{cell.data$data}). It can be obtained from \code{cell.load.alt} or \code{load_cell_data}.}

\item{filters}{A list of polygon dataframes with columns: x (values) y (values) xvar (variable name for x values) yvar (variable name for y values) type ("Include" or "Exclude")}

\item{unique_id_fields}{Vector with the names of the columns which uniquely identify each observation (a "primary key") for each data point (i.e. the "ucid" is not suficcient for time series datasets).}

\item{truthMode}{Priority for "Exclude" and "Include" polygon filter types, passed to \code{\link{calculateTruth}}. Must be either "all" (exclusion overcomes inclusion) or "any" (inclusion overcomes exclusion).}
}
\description{
Apply shiny-filters to cdata
}
