% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shiny.R
\name{calculateTruth}
\alias{calculateTruth}
\title{Calculate how filters apply to data according to the logic "mode"}
\usage{
calculateTruth(
  filterDF,
  cell_unique_id_field = "ucid",
  truth_column = "filter",
  mode = "all"
)
}
\arguments{
\item{filterDF}{The "cfilter" dataframe, as built by \code{\link{polyFilterApply}} by iteratively calling \code{\link{polyFilterCell}}.}

\item{truth_column}{Name of the column with the boolean filter.}

\item{mode}{Must be either "all" ("Subtractive" overrides "Additive") or "any" ("Additive" overrides "Subtractive").}
}
\description{
This function is meant to be called exclusively by \code{\link{polyFilterApply}},
as it produces appropriate inputs.
}
\keyword{internal}
